Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Audio volume is actually lower with this libaudio.  Also, there's no distortion because the volume is so low (which isn't exactly a "fix" for the distortion issue).  I'm going to say DO NOT MERGE because this needs a bit more research.

Patch-set: 1