Update patch set 4
Patch Set 4:

Thanks for pointing that out.  Looks like I was confused about the libril-*.so libraries.

It looks like I didn't break the radio because I got lucky; my vendor/huawei/u8150-vendor-blobs.mk is an old one which still has the stock libril.so to be copied.

I will experiment and resubmit this patch or abandon it.

Patch-set: 4