Update patch set 5
Patch Set 5:

Patch set 5 clarifies that what I was actually doing was removing the local libril (not "switching" to CM's libril).  I had misunderstood the makefiles.  We are still using the stock libril.so.

Tested on my U8150 device.  RIL still works as expected, but the device tree is now cleaner because of this.

Patch-set: 5