Update patch set 1
Patch Set 1:

Alan,

Did you try "downgrading" the copybit to the msm7k version? I tested rotation with this version for a while and the green screen didn't occur, but perhaps I wasn't thorough enough in testing.

Regardless, disabling the rotation animation is the way to go purely due to performance (and I've since made the change on my port too) :)

Patch-set: 1